Abstraction of AIState and fixed rest of scripts to work with the melee grunt refactor
