+++ Penteract/Assets/Scripts/AIMovement.h
@@ -2,9 +2,10 @@
 
 #include "Scripting/Script.h"
 
+#include "AIState.h"
+
 class ComponentTransform;
 class ComponentAgent;
-class AIMeleeGrunt;
 
 class AIMovement : public Script
 {
@@ -15,7 +16,10 @@ public:
 	void Start() override;
 	void Update() override;	
 
-	void Seek(const AIState state, const float3& newPosition, int speed);
+	void Seek(AIState state, const float3& newPosition, int speed);
+
+	bool CharacterInSight(const GameObject* character, const float searchRadius);
+	bool CharacterInMeleeRange(const GameObject* character, const float meleeRange);
 	
 
 private:
