+++ Penteract/Assets/Scripts/AIState.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum class AIState {
+	START,
+	SPAWN,
+	IDLE,
+	RUN,
+	ATTACK,
+	DEATH
+};
