+++ Penteract/Assets/Scripts/FangMovement.cpp
@@ -3,7 +3,7 @@
 #include "GameObject.h"
 #include "GameplaySystems.h"
 #include "Resources/ResourceMaterial.h"
-#include "AIMovement.h"
+#include "AIMeleeGrunt.h"
 
 EXPOSE_MEMBERS(FangMovement) {
 	// Add members here to expose them to the engine. Example:
@@ -47,7 +47,7 @@ void FangMovement::Update() {
 					ResourceMaterial* material = GameplaySystems::GetResource<ResourceMaterial>(meshRenderer->materialId);
 					material->diffuseColor = float4(Colors::Red(), 1.0);
 				}
-				AIMovement* enemyScript = static_cast<AIMovement*>(hitGo->GetComponent<ComponentScript>()->GetScriptInstance());
+				AIMeleeGrunt* enemyScript = static_cast<AIMeleeGrunt*>(hitGo->GetComponent<ComponentScript>()->GetScriptInstance());
 				enemyScript->HitDetected();
 
 			} else {
